# Optimize graph partition