--- a/include/onnxruntime/core/framework/tensor.h
+++ b/include/onnxruntime/core/framework/tensor.h
@@ -164,8 +164,16 @@ class Tensor final {
   /**
   The number of bytes of data.
   */
-  size_t Size() const noexcept {
-    return shape_.Size() * dtype_->Size();
+  size_t Size() const {
+    size_t ret;
+    int64_t l = shape_.Size();
+    if (l >= static_cast<int64_t>(std::numeric_limits<ptrdiff_t>::max())) {
+      ORT_THROW("tensor size overflow");
+    }
+    if (!IAllocator::CalcMemSizeForArray(static_cast<size_t>(shape_.Size()), dtype_->Size(), &ret)) {
+      ORT_THROW("tensor size overflow");
+    }
+    return ret;
   }
 
   // More API methods.