--- a/onnxruntime/core/framework/graph_partitioner.cc
+++ b/onnxruntime/core/framework/graph_partitioner.cc
@@ -55,6 +55,61 @@ KernelDefBuilder& BuildFusedKernelDef(KernelDefBuilder& builder, const onnxrunti
   return builder;
 }
 
+/**
+ * Check if a node can be placed on a specific provider.
+ * Do nothing if the node is already assigned
+ * \param graph
+ * \param capability
+ * \param kernel_registry_mgr
+ * \param provider_type name of the provider to test
+ * \param count A counter for generating fused node names. Should be unique within this subgraph
+ * \return Fused node. Return nullptr if there is no fuse
+ */
+static Node* PlaceNode(Graph& graph, std::unique_ptr<IndexedSubGraph> capability,
+                       const KernelRegistryManager& kernel_registry_mgr, const std::string& provider_type, int& count) {
+  if (nullptr == capability) {
+    return nullptr;
+  }
+
+  if (nullptr == capability->GetMetaDef()) {
+    // The <provider> can run a single node in the <graph> if not using meta-defs.
+    // A fused kernel is not supported in this case.
+    ORT_ENFORCE(1 == capability->nodes.size());
+
+    auto node = graph.GetNode(capability->nodes[0]);
+    if (nullptr != node && node->GetExecutionProviderType().empty()) {
+      // The node was not fused or assigned. Assign it to this <provider>.
+      node->SetExecutionProviderType(provider_type);
+    }
+  } else {
+    // The <provider> can run a fused <sub_graph> in the <graph>.
+    ORT_ENFORCE(nullptr != capability->GetMetaDef());
+    // Check whether any node in the <sub_graph> was already assigned.
+    bool sub_graph_available_for_assignment = true;
+    for (auto node_index : capability->nodes) {
+      auto node = graph.GetNode(node_index);
+      if (nullptr == node || !node->GetExecutionProviderType().empty()) {
+        // The node was fused or assigned, so that the whole sub-graph will not be assigned to this <provider>
+        // The assumption is that this <provider> can only run the sub-graph as a whole unit.
+        sub_graph_available_for_assignment = false;
+        break;
+      }
+    }
+    if (sub_graph_available_for_assignment) {
+      std::ostringstream oss;
+      oss << provider_type << "_" << capability->GetMetaDef()->name << "_" << count++;
+      std::string node_name = oss.str();
+      auto& fused_node = graph.FuseSubGraph(std::move(capability), node_name);
+      fused_node.SetExecutionProviderType(provider_type);
+      // searching in kernel registries, if no kernel registered for the fused_node, use compile approach
+      if (!kernel_registry_mgr.HasImplementationOf(fused_node, provider_type)) {
+        return &fused_node;
+      }
+    }
+  }
+  return nullptr;
+}
+
 Status GraphPartitioner::Partition(Graph& graph, bool export_dll, FuncManager& func_mgr) const {
   // It is a greedy partitioning algorithm per provider preferences user provided when calling ONNX RUNTIME right now.
   // 1. Execution providers' capabilities are checked one by one.
@@ -62,7 +117,8 @@ Status GraphPartitioner::Partition(Graph& graph, bool export_dll, FuncManager& f
   //    NOTE: A 'sub-graph' is a subset of nodes within the current Graph instance.
   //          The control flow nodes have nested Graph instance/s which are also called subgraphs,
   //          but are completely separate Graph instances and not a subset of nodes within a single Graph instance.
-  // 3. CPU execution provider is expected to be able to run any node and is the last one in execution provider preference.
+  // 3. CPU execution provider is expected to be able to run any node and is the last one in execution provider
+  //    preference.
   if (providers_.Empty()) {
     return Status(ONNXRUNTIME, INVALID_ARGUMENT, "No provider specified.");
   }
@@ -80,13 +136,7 @@ Status GraphPartitioner::Partition(Graph& graph, bool export_dll, FuncManager& f
   // It is only visible for current session.
   std::shared_ptr<KernelRegistry> fused_kernel_registry = std::make_shared<KernelRegistry>();
   // Partitioning <graph> based on provider preference and their capabilities.
-  auto kernel_registries = kernel_registry_mgr_.GetAllKernelRegistries();
-
-  std::vector<std::vector<std::unique_ptr<ComputeCapability>>> capabilities_of_all_providers;
   GraphViewer graph_viewer(graph);
-  for (auto& provider : providers_) {
-    capabilities_of_all_providers.push_back(provider->GetCapability(graph_viewer, kernel_registries));
-  }
 
   // If an execution provider return the capability that he could run a sub-graph,
   // onnxruntime will fuse the sub-graph into a function node. if the execution provider
@@ -98,59 +148,19 @@ Status GraphPartitioner::Partition(Graph& graph, bool export_dll, FuncManager& f
   // TODO: when the graph contain a function node, and user pass in the dll which could
   // run the function by SessionOption, we should create a function kernel for it and
   // delegate the compute to the functions inside the dlls.
-  int i = 0;
   for (auto& provider : providers_) {
     int count = 0;
     std::vector<Node*> nodes_need_compile;
-    for (auto& capability : capabilities_of_all_providers[i++]) {
-      if (nullptr == capability || nullptr == capability->sub_graph) {
-        continue;
-      }
-
-      if (nullptr == capability->sub_graph->GetMetaDef()) {
-        // The <provider> can run a single node in the <graph> if not using meta-defs.
-        // A fused kernel is not supported in this case.
-        ORT_ENFORCE(1 == capability->sub_graph->nodes.size());
-
-        auto node = graph.GetNode(capability->sub_graph->nodes[0]);
-        if (nullptr != node && node->GetExecutionProviderType().empty()) {
-          // The node was not fused or assigned. Assign it to this <provider>.
-          node->SetExecutionProviderType(provider->Type());
-        }
-      } else {
-        // The <provider> can run a fused <sub_graph> in the <graph>.
-        ORT_ENFORCE(nullptr != capability->sub_graph->GetMetaDef());
-        // Check whether any node in the <sub_graph> was already assigned.
-        bool sub_graph_available_for_assignment = true;
-        for (auto node_index : capability->sub_graph->nodes) {
-          auto node = graph.GetNode(node_index);
-          if (nullptr == node || !node->GetExecutionProviderType().empty()) {
-            // The node was fused or assigned, so that the whole sub-graph will not be assigned to this <provider>
-            // The assumption is that this <provider> can only run the sub-graph as a whole unit.
-            sub_graph_available_for_assignment = false;
-            break;
-          }
-        }
-        if (sub_graph_available_for_assignment) {
-          std::string node_name = provider->Type() + "_" + capability->sub_graph->GetMetaDef()->name + "_" + std::to_string(count++);
-          auto& fused_node = graph.FuseSubGraph(std::move(capability->sub_graph), node_name);
-          fused_node.SetExecutionProviderType(provider->Type());
-          // searching in kernel registries, if no kernel registered for the fused_node, use compile approach
-          bool need_compile = true;
-          for (auto* kernel_registry : kernel_registries) {
-            if (kernel_registry->TryFindKernel(fused_node, provider->Type())) {
-              need_compile = false;
-              break;
-            }
-          }
-
-          if (need_compile)
-            nodes_need_compile.push_back(&fused_node);
-        }
+    std::vector<std::unique_ptr<ComputeCapability>> capabilities =
+        provider->GetCapability(graph_viewer, kernel_registry_mgr_.GetKernelRegistriesByProviderType(provider->Type()));
+    for (auto& capability : capabilities) {
+      Node* n = PlaceNode(graph, std::move(capability->sub_graph), kernel_registry_mgr_, provider->Type(), count);
+      if (n != nullptr) {
+        nodes_need_compile.push_back(n);
       }
     }
 
-    if (nodes_need_compile.size() > 0) {
+    if (!nodes_need_compile.empty()) {
       if (export_dll) {
         std::string dll_path;
         ORT_RETURN_IF_ERROR(provider->Compile(nodes_need_compile, dll_path));
@@ -159,20 +169,24 @@ Status GraphPartitioner::Partition(Graph& graph, bool export_dll, FuncManager& f
       } else {
         std::vector<NodeComputeInfo> node_compute_funcs;
         ORT_RETURN_IF_ERROR(provider->Compile(nodes_need_compile, node_compute_funcs));
-        ORT_ENFORCE(node_compute_funcs.size() == nodes_need_compile.size(), "Provider doesn't return correct number of compiled functions");
-        for (auto j = 0; j < nodes_need_compile.size(); j++)
-          ORT_RETURN_IF_ERROR(func_mgr.AddFuncInfo(nodes_need_compile[j]->Name(), node_compute_funcs[j].compute_func, node_compute_funcs[j].create_state_func, node_compute_funcs[j].release_state_func));
+        ORT_ENFORCE(node_compute_funcs.size() == nodes_need_compile.size(),
+                    "Provider doesn't return correct number of compiled functions");
+        for (size_t j = 0; j < nodes_need_compile.size(); j++)
+          ORT_RETURN_IF_ERROR(func_mgr.AddFuncInfo(nodes_need_compile[j]->Name(), node_compute_funcs[j].compute_func,
+                                                   node_compute_funcs[j].create_state_func,
+                                                   node_compute_funcs[j].release_state_func));
       }
       for (auto* node : nodes_need_compile) {
         //prepare the func kernel
         KernelDefBuilder builder;
         BuildFusedKernelDef(builder, *node);
-        fused_kernel_registry->Register(builder, [](const OpKernelInfo& info) { return new FunctionKernel(info); });
+        ORT_RETURN_IF_ERROR(fused_kernel_registry->Register(
+            builder, [](const OpKernelInfo& info) { return new FunctionKernel(info); }));
       }
     }
   }
 
-  ORT_ENFORCE(graph.Resolve().IsOK());
+  ORT_RETURN_IF_ERROR(graph.Resolve());
 
   // To see if the node with no provider can be inlined. If one such nodes can be
   // successfully inlined, we re-run the partitioner on the modified graph.
@@ -183,10 +197,9 @@ Status GraphPartitioner::Partition(Graph& graph, bool export_dll, FuncManager& f
       if (nullptr == node_func) {
         continue;
       }
-      Status inliner_status = graph.InlineFunction(node);
       // If the node has a functionbody with no kernel and cannot be inlined
       // it is a invalid function
-      if (!inliner_status.IsOK()) return inliner_status;
+      ORT_RETURN_IF_ERROR(graph.InlineFunction(node));
       // Set the flag for re-run graph partition after successful inlining
       inline_flag = true;
       break;
@@ -212,7 +225,7 @@ Status GraphPartitioner::Partition(Graph& graph, bool export_dll, FuncManager& f
   }
 #endif
 
-  kernel_registry_mgr_.RegisterKernelRegistry(fused_kernel_registry, KernelRegistryPriority::HighPriority);
+  if (!fused_kernel_registry->IsEmpty()) kernel_registry_mgr_.RegisterKernelRegistry(fused_kernel_registry);
 
   return Status::OK();
 }