… onnxruntime/core/framework/kernel_registry.cc
… onnxruntime/core/framework/kernel_registry_manager.cc
… onnxruntime/core/framework/kernel_registry_manager.h
… onnxruntime/core/framework/session_state_initializer.cc
… onnxruntime/core/graph/model.cc
… onnxruntime/core/providers/cpu/cpu_execution_provider.h
… onnxruntime/core/session/inference_session.cc
… onnxruntime/test/framework/allocation_planner_test.cc
… onnxruntime/test/framework/execution_frame_test.cc
… onnxruntime/test/framework/memcpy_transformer_test.cc
… onnxruntime/test/onnx/microbenchmark/model_init.cc
… onnxruntime/test/onnx/runner.cc