Before this change,
If there are 3 providers, GPU, CPU, mkldnn
Then for each node, KernelRegistry::TryFindKernel will be called 3x3 = 9 times, instead of 3.
N providers need O(N^2) searches. It should be O(N).

Perf result:

**Baseline**:

Benchmark                  |Time          | CPU |Iterations
------------------------|------------|-------|------------
BM_SearchKernelRegistry_SingleNodeModel            |245 ns        |244 ns  |172212113
BM_SearchKernelRegistry_RealModel_tiny_yolo       |9551 ns       |9540 ns    |4412595
BM_SearchKernelRegistry_RealModel_inception_v4       |149148 ns       |148933 ns    |281947
BM_PartitionModel_tiny_yolo       |1971998 ns    |1961591 ns       |1435
BM_PartitionModel_inception_v4    |4453991 ns    |4443086 ns        |626

New code:

Benchmark                                                                     |Time                |Time Diff   |  CPU |Iterations
----------------------------------------------------------- |----------------|-------|-------|-------------------
BM_SearchKernelRegistry_SingleNodeModel                |291 ns             | +19%|291 ns  |143936682
BM_SearchKernelRegistry_RealModel_tiny_yolo            |11401 ns         |+19% |11389 ns    |3681155
BM_SearchKernelRegistry_RealModel_inception_v4      |138977 ns       |-7%    | 138789 ns    |302019
BM_PartitionModel_tiny_yolo                                        |1686156 ns     |-16%  | 1676349 ns       |1673
BM_PartitionModel_inception_v4                                  |1683208 ns     |-62%  |1676687 ns       |1656



For each kernel search call, it's about 20 percent slower.
However, overall, the graph partition process is significantly faster. And the new implementation works much better on complex models.