@linkerzhang 2019-02-12
quick question: can we not regress the perf on small model (first 3 rows)? :)