@snnn 2019-02-12
> quick question: can we not regress the perf on small model (first 3 rows)? :)

The top 3 rows are for KernelRegistryManager::SearchKernelRegistry function. Whether it's small model or large model, it's always slower.

The other rows are for GraphPartioner ,which will call KernelRegistryManager::SearchKernelRegistry function many many times. Even KernelRegistryManager::SearchKernelRegistry is slower,  GraphPartioner is much faster.